Removing unnecessary `Loop` struct
